post

+++ src/main/java/com/ramenshop/controller/MenuApiController.java
@@ -1,13 +1,16 @@
 package com.ramenshop.controller;
 
 import java.io.File;
+import java.io.IOException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,7 +26,7 @@ public class MenuApiController {
 	MenuService menuService;
 
     @PostMapping("/admin/edit")
-    public String write(
+    public String postMenu(
     		@RequestParam(name="name") String name,
     		@RequestParam(name="price") int price,
     		@RequestParam(name="discription") String discription,
@@ -62,15 +65,36 @@ public class MenuApiController {
     }
     
     
+//    @DeleteMapping("/admin/edit")
+//    public String deleteMenu(@RequestBody Menu menu) {
+//    	Menu findMenu = menuService.findMenuId(menu.getId());
+//    	if(findMenu == null) {
+//    		return "삭제실패, 데이터 미존재";
+//    	} else {
+//    		menuService.deleteMenu(menu.getId());
+//    		return "삭제성공!";
+//    	}
+//    }
+    
     @DeleteMapping("/admin/edit")
-    public String write(@RequestParam(name="id") Long id, HttpServletResponse response) {
-    	try {
-    	menuService.deleteMenu(id);
-    	response.sendRedirect("/admin/list");
-    	return "삭제성공";
-    	} catch(Exception e) {
-    		e.printStackTrace();
-    		return "삭제실패";
+    public String deleteMenu(@RequestParam(value="id") Long id) {
+    	
+    	if(id== null) {
+    		System.out.println("올바르지 않은 접근");
+    		return "redirect:/admin/list";
+    	} try {
+    		boolean isDeleted = menuService.deleteMenu(id);
+    		if(isDeleted == false) {
+    			System.out.println("게시글 삭제 실패");
+    		}
+    	}catch(DataAccessException e) {
+    		System.out.println("데이터 처리과정 문제");
+    	}catch(Exception e) {
+    		System.out.println("시스템에 문제");
     	}
+    	
+    	return "redirect:/admin/list";
+    	
     }
+
 }
